Split Export Controller, introduce DispatcherAware trait
